Refactor module config into submodules
