+++ dve/config/__init__.py
@@ -1,0 +1,5 @@
+"""
+This module simplifies and abstracts out accessing values from the
+configuration. Not all config access uses these functions, but the gnarly
+cases are centralized here.
+"""
