+++ dve/config/text.py
@@ -1,156 +1,21 @@
-"""
-This module simplifies and abstracts out accessing values from the
-configuration. Not all config access uses these functions, but the gnarly
-cases are centralized here.
-"""
-
-import os.path
 import logging
-from pkg_resources import resource_filename
 from dash import html
 from dash import dcc
 import dash_bootstrap_components as dbc
-from dve.dict_utils import path_get
+
+from dve.config.values import units_suffix, dv_units, nice_units, dv_tier
 from dve.text_utils import interpret, card_set
 
 logger = logging.getLogger(__name__)
 
 
-def file_exists(filepath):
-    filepath = resource_filename("dve", filepath)
-    return os.path.isfile(filepath)
-
-
-def validate_filepath(filepath):
-    if not file_exists(filepath):
-        logger.warning(f"'{filepath}' does not exist or is not a file")
-
-
-def validate_config(config, cfg_path, log=logger.warning, **kwargs):
-    """
-    Validate the config addressed by `cfg_path`. If there is nothing at the
-    specified path in the config object, log this fact. Otherwise, validate
-    the filepath found at that path.
-    """
-    filepath = path_get(config, cfg_path, **kwargs)
-    if filepath is None:
-        log(f"Expected config at '{cfg_path}' is not present.")
-    else:
-        validate_filepath(filepath)
-
-
-def validate(config):
-    """
-    Validate a configuration. Specifically:
-    - Check that all filepaths specified in config exist. Log a warning message
-      for those which do not exist.
-    """
-    for filepath in config["values"]["paths"].values():
-        validate_filepath(filepath)
-
-    for design_variable in config["values"]["ui"]["dvs"]:
-
-        # Validate historical filepaths, if present
-        cfg_base_path = f"dvs/{design_variable}/historical/datasets"
-        if path_get(config, cfg_base_path, separator="/") is None:
-            logger.info(
-                f"No historical datasets specified for {design_variable}"
-            )
-        else:
-            for cfg_ext_path in (
-                "model",
-                "reconstruction",
-                "stations/path",
-                "table_C2",
-            ):
-                validate_config(
-                    config, f"{cfg_base_path}/{cfg_ext_path}", separator="/"
-                )
-
-        # Validate future filepaths, if present.
-        cfg_base_path = f"dvs/{design_variable}/future/datasets"
-        if path_get(config, cfg_base_path, separator="/") is None:
-            logger.info(f"No future datasets specified for {design_variable}")
-        else:
-            for cfg_ext_path in config["values"]["ui"]["future_change_factors"]:
-                validate_config(
-                    config, f"{cfg_base_path}/{cfg_ext_path}", separator="/"
-                )
-
-
-def filepath_for(
-    config,
-    design_variable,
-    climate_regime,
-    historical_dataset_id=None,
-    future_dataset_id=None,
-):
-    root_path = f"values/dvs/{design_variable}/{climate_regime}/datasets"
-
-    if climate_regime == "historical":
-        ext_path = {
-            "stations": "stations/path",
-            "table": "table_C2",
-            "model": "model",
-            "reconstruction": "reconstruction",
-        }[historical_dataset_id]
-    else:
-        ext_path = future_dataset_id
-
-    return path_get(config, f"{root_path}/{ext_path}", separator="/")
-
-
-def filepath_defined(*args, **kwargs):
-    return filepath_for(*args, **kwargs) is not None
-
-
-def dv_has_climate_regime(config, design_variable, climate_regime):
-    """
-    Return a boolean indicating whether a DV has definitions for specific
-    climate regime (historical or future) datasets.
-    """
-    return climate_regime in config["values"]["dvs"][design_variable]
-
-
 def dv_name(config, lang, design_variable):
     """
     Return the name of a design variable. Currently, this is the internal
-    id of the DV, so `config` and `lang` are not used. 
+    id of the DV, so `config` and `lang` are not used.
     That could conceivably change.
     """
     return design_variable
-
-
-def dv_tier(config, design_variable):
-    return config["values"]["dvs"][design_variable]["tier"]
-
-
-def nice_units(config, units):
-    try:
-        definition = config["values"]["units"][units]
-        return (
-            definition["nice"],
-            (" " if definition.get("separator", True) else ""),
-        )
-    except KeyError:
-        return units, " "
-
-
-def units_suffix(units, pattern=" ({units})"):
-    if units is None or units == "":
-        return ""
-    return pattern.format(units=units)
-
-
-def dv_units(config, design_variable, climate_regime, nice=True):
-    """
-    Return the units of a given design variable, for historical or future
-    projections.
-    """
-    units = config["values"]["dvs"][design_variable][climate_regime]["units"]
-    if not nice:
-        return units
-    return nice_units(config, units)[0]
 
 
 def dv_label(
@@ -175,45 +40,6 @@ def dv_label(
         else ""
     )
     return f"{dv_name(config, lang, design_variable)}{description}{units}"
-
-
-def dv_roundto(config, design_variable, climate_regime):
-    return config["values"]["dvs"][design_variable][climate_regime]["roundto"]
-
-
-def dv_colour_map(config, design_variable, climate_regime):
-    return config["values"]["dvs"][design_variable][climate_regime][
-        "colour_map"
-    ]
-
-
-def dv_colour_scale_default(config, design_variable, climate_regime):
-    return config["values"]["dvs"][design_variable][climate_regime]["scale"][
-        "default"
-    ]
-
-
-def dv_colour_scale_disable_logarithmic(
-    config, design_variable, climate_regime
-):
-    return config["values"]["dvs"][design_variable][climate_regime][
-        "scale"
-    ].get("disable_logarithmic", False)
-
-
-def dv_colour_bar_sigfigs(config, design_variable, climate_regime):
-    return config["values"]["dvs"][design_variable][climate_regime]["colorbar"][
-        "sigfigs"
-    ]
-
-
-def dv_historical_stations_column(config, design_variable):
-    return config["values"]["dvs"][design_variable]["historical"]["datasets"][
-        "stations"
-    ]["column"]
-
-
-# Text(ish)
 
 
 def app_title(config, lang):
